refactor: App 컴포넌트를 구조에 따라 다시 Header, Main, Footer 컴포넌트로 분리
