+++ src/components/FolderInfo.js
@@ -4,7 +4,7 @@ function FolderInfo({ folderName, ownerName, image }) {
   return (
     <div className="folderInfo">
       <img className="folderImg" src={image} alt="폴더 소유자 프로필 이미지" />
-      <p class="ownerName">{ownerName}</p>
+      <p className="ownerName">{ownerName}</p>
       <h1 className="folderName">{folderName}</h1>
     </div>
   );
